+++ app/page.tsx
@@ -1,7 +1,7 @@
-import Image from "next/image";
-
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24"></main>
+    <div className="min-h-screen pt-24 pb-16 px-6 md:px-16 lg:px-24 xl:px-32">
+      <h1 className="text-4xl font-bold text-foreground mb-6">Calculator</h1>
+    </div>
   );
 }